Moved HymnItem into HymnList
